Rework integration tests organization
