+++ tests/IcecatConnector/Integration/Job/EnrichProductsTest.php
@@ -3,7 +3,6 @@
 namespace Pim\Bundle\IcecatConnectorBundle\Tests\Job;
 
 use Akeneo\Bundle\BatchBundle\Command\BatchCommand;
-use Pim\Bundle\IcecatConnectorBundle\Tests\AbstractTestCase;
 use Pim\Component\Catalog\Model\ProductInterface;
 
 /**
@@ -11,14 +10,18 @@ use Pim\Component\Catalog\Model\ProductInterface;
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class EnrichProductsTest extends AbstractTestCase
+class EnrichProductsTest extends AbstractJobTestCase
 {
+    /** @var string */
     private $jobCode = 'icecat_enrich_products';
 
-    public function additionnalSetup()
+    /**
+     * {@inheritdoc}
+     */
+    protected function setUp()
     {
+        parent::setUp();
         $this->createImportProfile('Icecat', $this->jobCode);
-        $this->loadData();
     }
 
     public function testEanProductIsEnriched()
@@ -42,7 +45,6 @@ class EnrichProductsTest extends AbstractTestCase
         $product = $this->get('pim_catalog.builder.product')->createProduct();
         $this->get('pim_catalog.updater.product')->update($product, $dataProduct);
         $this->get('pim_catalog.saver.product')->save($product);
-
         sleep(2);
 
         copy(__DIR__ . '/../Resources/mapping.csv', '/tmp/mapping.csv');
